This is a fuller sketch of the HumanClientFSM, but it is still just a sketch. Most of it is quite badly broken for the time being.  Overall, only saving/ autosaving and starting a SP game are working again.  Specific changes include:
- Implemented all the react() methods of the FSM states.
- Cleaned up/modified several Message creation functions.
- Massive cleanup/reorg of the ClientUI.  All Screen*() methods are removed,
and the ClientUI only creates a single Wnd, the MapWnd.  All others are now
created, and all transition logic between them is now handled by, the FSM.
- TurnProgressWnd no longer creates/modifies/owns CombatWnd; now the FSM does
that.
- Lots of MapWnd grooming.
- IntroScreen was largely gutted.  All of it's application-level logic
(connecting to the server, sending messages, etc.) was moved to HumanClientApp.
- MultiplayerLobbyWnd was similarly gutted (app logic moved to app).
- HumanClientApp was massively reorganized.  It now delegates most of its
single-use code to the FSM, which exists purely as an implementation detail of
the HumanClientApp.  Its code which is used in multiple calling contexts is
retained in its member functions.

svn path=/branches/networking-reimplementation/FreeOrion/; revision=2142