Changed the CombatStartMessage to contain a full CombatData object.  Specific changes in support of this:
- CombatObject now serializes its base as
"SimpleVehicle". "OpenSteer::SimpleVehicle" was causing problems, since XML
tags cannot contain colons.
- Added a serialization class export for Missile.
- Added code to CombatData's ctor to add ships to the pathing engine.
- Sketched in the AI interfaces for handling combat setup (pre-combat unit
placement, etc.).
- Changed ProximityDatabase's serialization code to only serialize the objects
it contains, not all the grid cells it contains.

Other changes:
- Changed NotEmpireFlag() to EnemyOfEmpireFlags(), and added the note there
that enemy determination is a TODO.  This replaces the numerous notes about
this throughout the combat classes.
- CombatShip::GetShip() now returns a reference instead of a pointer,
reflecting the invariant that the return value must exist.
- Added CombatEventListener inheritance to CombatWnd.  Most of the interface
is unimplemented so far.
- Changed the way CombatWnd accounts for resources related to ships, and based
on the changes, implemented CombatWnd::RemoveShip().

svn path=/trunk/FreeOrion/; revision=2968