Condition / ValueRef / Effect Adjustments
-Made the stored types of Conditions, Effects, and ValueRefs non-const
-Added a special-case in ValueRef::Constant<std::string> that, if the
value is "CurrentContent", the value of a new member variable
m_top_level_content is returned instead of the value itself.
-Added SetTopLevelContent functions to ValueRef, Condition, and Effect,
which is called recursively to any contained member of any of these
types, until a ValueRef::Constant is reached, which stores the value in
a new data member used for the previous point.
-When a ValueRef, Condition, or EffectsGroup is added to some content,
such as a BuildingType, Special, Species, etc, the SetTopLevelContent
function is called with the content's name, so that the value can be
returned by any ValueRef::Constant<std::string> within.
-Parsing for CurrentContent without quotes as a
ValueRef::Constant<std::string>.
-Grooming / safety checks