Fix resize tab on graphical combat summary.
Cause CombatReportWnd to respond to mouse events over the part of its
resize tab that is covered by the GraphicalSummary.  Confusingly, the
resize tab was drawn _over_ the GraphicalSummary but the latter was still
receiving the mouse events.

Also make MapWnd hide the combat summary window on MapWnd::Cleanup(). This
prevents the combat summary window form sticking around if it's open
when the player resigns from a game.