Add size constraints for combat report window.
CombatReportWnd now checks for minimum size (determined by the contents of
the window) and maximum size (determined by the size of the app window) on
SizeMove(...).

Add functions to BarSizer, OptionsBar, GraphicalSummaryWnd and
CombatReportPrivate to calculate the minimum size of CombatReportWnd.

Add minimum width for participant bars to prevent low-health ones
disappearing when TOGGLE_BAR_WIDTH_PROPORTIONAL is activated.