Remove max size checks from combat report window.
Remove connection to WindowResizedSignal and references to AppWidth &
AppHeight that were added to this window only, as that type of functionality
belongs in a superclass if it is added.

It is possible for the combat report to become larger than the app window
when using certain settings in rare circumstances but the user can recover
from this condition.