Fix SidePanel::InWindow giving false positives.
When m_system_resource_summary is empty, it is made invisible and moved to
the left side of the app window but SidePanel::InWindow would still return
true and cause the SidePanel to respond to mouse events in that area (even
though it is visually on the other side of the app window).  Adding a
check that m_system_resource_summary is non-empty prevents this.

Also fix whitespace errors.