Add minimum size check for combat report window.
CombatReportWnd::SizeMove now respects the minimum size of its currently
open tab and GraphicalSummaryWnd::MinUsableSize is now implemented for
this purpose, along with various functions in its children and helpers.

Did not use GG::Wnd::SetMinSize because that makes it difficult to force a
maximum size via polymorphism, which is planned for CUIWnd-derived
classes.