Add size & position validation for CUIWnd::SizeMove
CUIWnd::SizeMove now makes sure that the CUIWnd stays within the bounds
of its parent (or the app window if it is parentless).  Added
CUIWnd::ValidatePosition() as a neater way to call SizeMove() with the
current position and size to trigger the position checking.  Removed
similar position-checking code from CUIWnd::LDrag.

MapWnd::DoLayout() now calls ValidatePosition() on the various sub-windows
that it can show, which updates their positions if the app window changes
size.