Don't reply on the python shebang during the CMake build
When invoking the make_versioncpp.py script we relied on the fact, that
`python` points to a python2.7 executable.  This assumption isn't
nessecarily true anymore for some linux distribution.  So instead on
relying the shebang line within the script we now use CMake to locate
the correct interpreter and explicit call the `make_versioncpp.py`
script with that interpreter.