Universe generation: additional requirement for home system selection
Home system selection process partially rewritten to implement another requirement home systems have to meet. Now not only a certain minimum of planets have to be in the near vicinity of a home system, but also a certain limit of systems. This is supposed to prevent home systems to be placed at dead ends with too few systems that can be reached easily (e.g. on the spiral arm maps).