Make OptionsDB keep options that it doesn't recognise (ver 2)
Add a new bool OptionsDB::Option::recognized to mark any options read from
an XML document that have not yet been registered using Add().  This
allows the DB to preserve information for options that are added
dynamically but functions such as Get(), OptionExists() etc. will still
throw until those options are registered with Add().

Add a private overload of OptionsDB::OptionExists that takes a map
iterator rather than a string because the code to determine whether an
option "exists" is now slightly more complicated and many OptionsDB
functions test for existence using iterators rather than the name of the
option.