Mark util/Version.cpp as autogenerated in CMake
Some users reported issues about a undefined reference to the
`FreeOrionVersionString` symbol when doing a build of a clean check out
of the FreeOrion code.  The implementation of this symbol is located
within the autogenerated util/Version.cpp, which is generated properly,
but wasn't compiled by the freeorioncommon target.  By marking the file
as autogenerated the freeorioncommon target now considers the file.

Fixes #184