Keep unrecognised options from the command line
Modify OptionsDB::SetFromCommandLine() to also store unrecognised options.
NB: options specified with short names (single dash) are not yet supported
for this, use their full names instead!

Improve differentiation between flag options and value options.  If an
option looks like a flag in the config (e.g. empty XML tag) but is later
registered with Add() an error message will be logged and it will fall
back on the default value specified in the call to Add().  If an option
looks like it has a value but is later registered with AddFlag() an error
message will also be logged but the flag will still be set to "true" and
the specified value will be ignored.

Fix a typo.