Modify make_versioncpp.py to allow nicer-looking version strings
Add a template keyword FreeOrion_BRANCH_ (note trailing underscore) which
is the same as FreeOrion_BRANCH but is given a trailing space character if
the string is otherwise not empty.

Modify Version.cpp.in and Info.plist.in to use FreeOrion_BRANCH_ so that
version strings generated for them don't have double-spaces if built on
branch "master" (because the branch name "master" is replaced with an
empty string in make_versioncpp.py) but otherwise still has a space on
either side of the branch name.

FreeOrion_Install_Script.nsi.in is also referenced in make_versioncpp.py
but does not use the FreeOrion_BRANCH keyword in the first place.