Make sorting stable in the object window .
Stable sorting means that if two objects in the list
are equal (with respect to the current sort key),
their order won't change. This is useful
for example if you sort first by planet size, then by owner,
the planets of each empire will be shown in size order
with respect to each other.

The stability is achieved by not rebuilding the list
from scratch with Refresh every time the sort key is changed,
the list box is smart enough to resort itself (and does it stably).
As the comment in the code states, nesting and sorting don't
mix. That is not a result of this change, but a pre-existing flaw.