Add HumanClientApp::FullscreenSwitchSignal...
...and connect to it in CUIWnd ctor.

Add a signal that is sent from HumanClientApp::Reinitialize() when it
detects a fullscreen/windowed transition.  Can't rely on
OptionsDB::OptionChangedSignal("fullscreen") because that can change
without actually triggering a video mode change (e.g. user still has to
click the apply button in the OptionsWnd) and
HumanClientApp::HandleWindowResize() doesn't provide video-mode-change
info.

Connect this signal to CUIWnd::LoadOptions() in the CUIWnd ctor if
config_name is specified so that it automatically loads the new options
when switching between fullscreen and windowed modes.