Add OptionsWnd widget to reset window positions...
...to their defaults.  While checked, the
option will cause all current CUIWnds and any new ones to reset their
saved values to the defaults (the defaults are calculated when the window
is constructed).

Change MapWnd ctor to use the default dimensions of the SitRepPanel to
calculate defaults for EncyclopediaDetailPanel, instead of the current
position of the SitRepPanel.