Add option to hide or keep open the map panels...
...while the production window is open.  Show/HideProduction() now calls
RemoveWindows() and RegisterWindows() instead of calling Show()/Hide() on
the map view panels (SitRepPanel, ObjectListWnd etc.) if the relevant
option is enabled in the OptionsWnd (in the Galaxy Map tab), which allows
those panels to remember their state.

Change MapWnd::Toggle[window]() functions to call their Show[window]()
counterparts if the production/research/design windows are open, instead
of simply hiding the latter windows.  This ensures that if e.g. a turn is
advanced with both the production window and sitrep panel open and the
sitrep toolbar icon is clicked, the production window is hidden and the
sitrep panel is updated with the latest turn's sitreps, instead of only
hiding the production window and leaving the sitrep panel un-updated.

When any of the panels' toolbar icons are clicked, the
production/research/design windows are closed and all of the panels are
restored.

Add calls to ShowAllPopups() and RemoveWindows() to MapWnd::Cleanup() to
ensure that popups don't save ".visible = false" to the config and that
windows are properly hidden when entering/leaving a game.

Bind panel close buttons to MapWnd::Hide[wnd]() rather than
::Toggle[wnd]().  Clicking the close button on a panel while the
production window is open now closes the panel instead of closing the
production window.