Revert commit 3323243
As "git show" returns the committer date as a local timestamp of the timezone of the committer, we can't use this for the build number. With that behavior it's possible that a later commit will produce a "lower" build number, which is not what we want. Reverting to the procedure which converts the timestamp to UTC before using it for the build number.