Remove (most of) the extra space in ListBoxes after the end of the content.
This was intentional behavior, so that any of the rows in the ListBox
could be in any position in the scroller, but I don't see the advantages
of that, and even if it is useful I think the dead space is ugly. There
still will often be a little padding after the last entry, so that it's
possible to scroll down far enough that the last row doesn't get cut off.