Move universe data tables into Python.
There's no point in loading the information from a text file and sending
it from C++ to Python, when it can just be in the Python files to begin
with.

I didn't port over the planet type adjustments since

* They weren't used;
* Most of them were all 0s
* I don't see any point in having a type adjustment; it just arbitrarily
  favours some species with certain habitabilities.

Of course, if someone wants to add them back it will just be a matter of
adjusting the Python code now.

I've only tested compilation on MSVC2013, so someone checking if the
changes to the other build processes worked would be appreciated.