Remove old, unused code related to empires losing/winning.
Most of this assumes that players are disconnected from the server when
they lose. There's been discussion of allowing empires to be restored to
the game via diplomacy, and it certainly should be possible to keep
playing after you get, say, a tech victory.

There was also code to send a Message to players when an empire lost/won,
but as far as things are concerned now it seems fine to just use sitrep
entries. Perhaps at some point, a special congratulatory screen would
be used upon victory?