Create test infrastructure for testing GG UI components
This commit adds a test application to test changes in the GG UI with
minimal effort.

You can create an application simply by creating and running an instance of
TestApp.  There is also a TestDialog that will help in testing how windows work
inside resizable dialogs.

Included is an example test that tests the GG text control.

Rebases and fixes #259