Create infrastructure for experimental applications.
This creates a folder, helper application and
CMake rules for making small binaries that
build quickly and cna be used to experiment with specific
pieces of the code base with minimal cycles.

You can create an application simply by creating and
running an instance of TestApp. There is also a TestDialog
that will help in testing how windows work inside resizable dialogs.

Included is an example test that tests the GG text control.