Change CUIWnd fxns to invalidate rather than...
...remove options.  RemoveUnusedOptions and RemoveWindowOptions were
renamed to Invalidate[...].  InvalidateWindowOptions() now sets the top
and left of the window to INVALID_POS for the current fullscreen/windowed
mode only.

This makes "resetting" certain windows (FleetWnd, OptionsWnd...) when they
are not open behave more like the rest of the windows: it only changes the
position for the current fullscreen/windowed mode.  Note that specifying
the "--window-reset" flag will still reset ALL window positions because
it's checked before any window options are registered.