Fix for Spatial Distortion Generator
This fixes issue#295. The problem was caused by a faulty special case handling in WithinStarlaneJumpsSimpleMatch for jumps = 0, which doesn't correctly handle fleets in transit between systems. As the reason for the special case handling mentioned in a comment does not apply (anymore?), I just removed the special case handling, which (together with an adjustment of the scope condition of the distortion focus in species.txt) fixed the problem.

I also adjusted the associated MoveTo effect so that not always the same neighbor system gets selected as destination for the "teleport" of the affected fleets.