Improve the behaviour of left-clicking on build queue entries.
Instead of just jumping the map to the system where the planet is located,
open up that system in the sidepanel and select the planet. This should
be especially handy in late game, when double-clicking on a system can
get to be very slow and it helps to do everything in one step.