Make a MultiEdit without a scroll bar pass mouse wheel events
Before this change, a MultiEdit would always swallow
mouse wheel events, even if it did not have any scrolling.
This was problematic when using a MultiEdit as a child of
a window that has its own scrolling.

Now a MultiEdit will pass scroll events to its parent
whenever it has no scrolling of its own.