Universe generation: added Python class to manage galaxy setup data
Added PyGalaxySetupData class on the Python side to store a copy of the galaxy setup data passed to the universe generation Python script. This way the universe generation process can modify/adjust these parameters if needed without having to modify the original data in the data structure provided by the FreeOrion Python interface.