Attempt to fix build error on Linux
Apparently the boost iostreams library doesn't get linked against the FO binaries on Linux, causing link errors. Most likely because the cake files didn't get adjusted to the recent addition of compressed xml serialization, which requires boost iostreams with lib support enabled. This commit is an attempt to add the required adjustments to the cmake files. As I'm not on Linux, I can't test it, so further adjustments might be necessary.