Revised implementation of random galaxy setup option picking
Using a hash of the seed and the number of available options to determine the option to pick for "random" selections produced the same pick for all settings that had the same number of options to pick from, which isn't sufficiently random.

This alternative approach extends the GalaxySetupData structure with additional private m_picked_* fields for all settings that offer a "random" option, and a member function that executes the random picking for all those settings (which gets called at the start of universe generation, right after the RNG has been initialized with the seed). The random picks are stored in the corresponding m_picked_* fields, and upon retrieval of an option the corresponding accessor function checks if the "random" option is set, in which case the value of the corresponding m_picked_* field is returned.