Universe generation: fixed non-reproducible distribution of space monsters even for the same seed
In the generate_monsters function monster fleet plans have been stored in and randomly selected from a dict, which resulted in different space monster distribution even for the same RNG seed. Fixed that by using a list to store the monster fleet plans in addition to the dict which is still required to keep track of how often a monster fleet has already been spawned.