Add some support for freedesktop XDG Base Directory Specification
The XDG Base Directory Specification is at
https://specifications.freedesktop.org/basedir-spec/basedir-spec-
latest.html

It requires that configuration files be stored in
$XDG_CONFIG_HOME/freeorion which defaults to
$HOME/.local/share/freeorion

It requires that user app data be stored in
$XDG_DATA_HOME/freeorion which defaults to
$HOME/.config/freeorion

This commit splits GetUserDir() into GetUserConfigDir() and
GetUserDataDir().  On linux and BSD they return the correct XDG
directories.  On Windows and OSX they return the same directory
as GetUserDir() before this commit.

This commit does not support XDG_DATA_DIRS, XDG_CONFIG_DIRS,
XDG_CACHE_HOME or XDG_RUNTIME_DIRS.