Create TextControl::MinUsableSize(X width)
TextControl::MinUsableSize(width) formats its text in order to determine
the actual minimum size occupied by the control. This is used in Layout
to correctly calculate the minimum height of a TextControl, assuming it
spans the entire width of the layout.

It caches previous results and only updates the result when the
requested width changes by more than one space character width.