Fixed setting planet type of empire homeworlds
When setting the planet type of an empire homeworld to the preferred type of the starting species in UniverseGenerator::SetEmpireHomeworld during universe generation, only the current type was set, but not the original type. Obviously that's not exactly the intention here, so setting the original type to the preferred type too has been added.