Gave Trith new Telepathic Detection species trait
Added new "Telepathic Detection" species trait which grants basic visibility of populated planets within a certain number of jumps of a ship or planet of the species with this trait and added the new trait to species Trith. Changed their Great Detection species trait to Good Detection for balance reasons.