Add python test framework using pytest with tests for Character.py
The framework is simply a README file with examples of how to run
pytest.

The test for Character.py test that all of the combiners for multiple
behaviors work as expected. The combiners are min, max, preference and
all True.

It tests that indexing by behavior selects the correct behavior.
It tests that building a Character out of non-behaviors raises a
TypeError.

The tests for CharacterStrings check that make_aggression table make a
table of the correct length that is indexed by the aggression behavior
of a Character that contains Aggression and show UNKNOWN_VALUE for
Characters that don't contain aggression.