Add debugging bypass to randomized Character and Behavior.
Add a debugging bypass to force individual AIs to have a specific
behavior or to force all AIs to have the same behavior. If some AIs have
an individual behavior forced and some do not, then the remainder use
the .all setting.

Add three types of options to OptionsDB to force the value of aggression
and empire id behaviors.  The .force option is a bool indicating whether
the bypass is active.  The .all option is a default if a specific AI
does not have its own value.  The .AI_<number> option forces that
behavior to that value for AI_<number>.

The options are:
AI.config.behavior.<behavior name>.force  -- If true use the following
options to bypass the behavior
AI.config.behavior.<behavior name>.all    -- If present use this value
for all of the AIs not individually set
AI.config.behavior.<behavior name>.AI_1   -- Use for AI_1
   .
   .
   .
AI.config.behavior.<behavior name>.AI_40  -- Use for AI_40

, where <behavior name> is either aggression or empire-id.

An example of the code from persistent_config.xml to setup some AIs
follows.

<AI>
    <config>
      <behavior>
        <aggression>
          <force>1</force>
          <all>4</all>
          <AI_0>5</AI_0>
          <AI_1>4</AI_1>
          <AI_2>3</AI_2>
          <AI_3>2</AI_3>
          <AI_4>1</AI_4>
          <AI_5>0</AI_5>
        </aggression>
        <empire-id>
          <force>1</force>
          <AI_0>5</AI_0>
          <AI_1>4</AI_1>
          <AI_2>3</AI_2>
          <AI_3>2</AI_3>
          <AI_4>1</AI_4>
          <AI_5>0</AI_5>
        </empire-id>
      </behavior>
    </config>
</AI>