Adding Character, Behavior and CharacterString to AI.
Create AI Character composed of Behaviors to encompass variation between
AIs.

Character solves the problem of having AI personalization added
piecemeal throughout the AI code.  Now each type of personalization is
encapsulated in a single Behavior.  The Behaviors are combined to form a
single unique Character per AI.

CharacterStrings are tables of strings, typically UserStrings indexed by
Character.  They return single strings or lists of strings to be used to
communicate in a Character appropriate manner with the human player.