Refactor specials.py to improve performance speed.
Apply base_chance to specials up front to remove objects that will not
have specials from the pool.

Define a GALAXY_DECOUPLING_DISTANCE which is the largest distance over
which specials interact with anything.

Consider groups of objects that are mutually farther apart than
GALAXY_DECOUPLING_DISTANCE so that their spawning rates can be
calculated collectively and quickly.