Refactor the GraphImpl portion of Universe into Pathfinder
Separate the functionality encompassed by the graph of starlanes, and
pathfinding into its own class.

The benefits will include the following.

Better code partitioning and a smaller Universe class.

Optimized neighborhood region finding functions, which are currently
prevented because the distance_matrix_storage and the implementation
of the distance cache are on the different sides of the public private
divide.  These functions are used directly in at least galaxy creation
and colonization. They are used indirectly in the costly condition
WithinStarlaneJumps.

Faster partial compile times since Pathfinder will be a pimpl
implementation, so that the distance_matrix_storage and cache are
implemented together.

The refactoring will be completed in steps.
1. * Separate the pathfinding functionality into Pathfinder.
2. Pimpl Pathfinder.
3. Refactor the distance_matrix_cache to expose rows for reading.
4. Implement the neighbor functions.
5. Use the the neighbor functions