Add examine_row() to distance_matrix_cache in Pathfinder.cpp
examine_row() is called with a row number ii, a cache hit function and a
cache miss function.  On a cache hit the entire row is passed to the cache
hit function.  On a miss the cache miss function is called to fill the row
and then the result is passed to the cache hit function.