Make hull stealth and detection descriptions agree with game mechanics.
The adjective used to describe ship hull stealth and detection were not
consistent with the numbers for game mechanics or other descriptions.

The adjectives used were: very low, low, poor, average, moderate,
normal, high, good, and excellent.

I reduced the used adjectives to: very low, low, medium, high and very
high.

The tables show values for each value of stealth or detection the number
of hulls, the adjective and the type of empire scanner capable of
detecting that level of stealth:

Stealth Value: -55 -35 |-15 -5  | (0)  5  |   15 25    |   35 45
       Number:  1   1  |  1  2  |  0  17  |    2  3    |   4  2
    Adjective:  v.low  |  low   | medium  |    high    |   v.high
      Scanner:         Basic (10)         | Active(30) | Neutron (50)

Detection Value: 10  |   25   |  50  | 75 100
         Number:  1  |   25   |   4  |  2   1
      Adjective: low | medium | high | v.high

I chose medium to be the most typical stealth or detection value.
I chose high and very high stealth to correspond to the thresholds for
the active and neutron scanners.
Note, negative stealth is not displayed in game, but it affects results.