In CombatLogWnd replace LinkText with LazyScrollerLinkText.
The expansion of LinkText by Font::DetermineLines is too slow to run
without freezing the UI.

LazyScrollerLinkText populates the text with and ellipsis and connects
to the scroll signal.  When it scrolls into view it replaces the text
with the real text and Font::DetermineLines only needs to process the
LinkTexts that are in view at one time.