Refactor JumpDistanceBetweenObjects
JumpDistanceBetweenObjects detected whether objects were not in a
system, in a system, ship or fleets in transit between two systems, for
each of two universe objects and then calculates the jump distance for
the correct case of the four possible cases: two systems, system and
fleet, fleet and system or fleet and fleet.

This commit creates ObjectLocation() which returns a boost::variant of
the single location or in transit locations.  This is passed to visitors
which only examine those cases.