Adjust JumpDistanceBetweenObjects when fleet path is lateral.
JumpDistanceBetweenObjects returns (max(jumps1, jumps2) - 1) for
travelling fleets which is correct if the fleet is travelling towards to
away from the candidate location.  If both the start and end system of a
travelling ship are at the same distance the from the other object then
the distance is short by one step.  If the start or end location of the
travelling fleet is the same as the first object's location then the
returned distance is -1.

This commit first checks that each end of the path is not -1 the no path
distance and then returns the minimum of the two distances, which is
correct when the start and end path are at the same distance from the
first system.