Change GenerateFleetName from static to non-static member function.
GenerateFleetName() was being passed a vector of ship ids and a system
id in order to generate a fleet name.  In all cases except when ordering
fleet creation the fleet already has the necessary information.

By interpreting a blank name as a request for a generated name it is
possible to change all the cases to use a non-static member function
GenerateFleetName().