Add Dismiss Sensor Ghost to FleetWnd context menu.
The UI displays known objects that have been detected in previous turns
as partially greyed out.

Sometimes these fleets disappear/merge/die outside of the view of the
player empire.  Then the ghost fleets are displayed in their last known
state forever, even though the player is fairly certain that the fleet
no longer exists and is very certain that they want the detected fleet
removed from the display.

This commit adds a context menu option "Dismiss Sensor Ghost" to ghost
fleets.  The option removes the fleet from the table of known objects.
If the fleet is detected again it is re-added to the table of known
objects as normal.