Extend test InitStarLaneRendering to draw orange triangles for debug.
When comparing old and new methods of rendering the supply network,
draws orange triangles to indicate systems the new method has
added/removed from the display.

Triangles below/above the system with an apex at the top/bottom indicate
the system added to/absent from the new rendering method relative to the
old.