Prevent Experimentor spawn location from sundering the galaxy.
Sometimes the experimentor system spawns in a location that divides the
rest of the galaxy into two or more disconnected pieces.

This commit prevents proposed spawning location that would sunder the
galaxy from being used.

It does this by computing the disjoint set of the galaxy without the
starlanes from the proposed system and checking that it is the correct
number of pieces.