Add CanAlterStarlanes tag to MonsterFleetPlan.
This commit add CanAlterStarlanes to the FOCS MonsterFleetPlanParser,
CanAlterStarlanes() to MonsterFleetPlan and can_alter_starlanes() to the
python wrapped MonsterFleetPlan.

CanAlterStarlanes() indicates that a monster fleet can alter starlanes.

This removes the local secret setting monsters_that_alter_starlanes from
monsters.py in favor of the more central and better localized location
in the monster definition.