Add SectionedScopedTimer.
SectionedScopedTimer works like ScopedTimer and it has a member function
EnterSection(name) which starts a new sub section timer.

When SectionedScopedTimer is destroyer it prints a table of the elapsed
time of each section and the whole timer.

Only one subsection timer is active at a time.

Subsection can be re-entered.