Replace boost::timer with boost::chrono in ScopedTimer.
Change ScopedTimer from using deprecated boost::timer to using
boost::chrono.

The deprecated  boost::timer was a wall clock on Windows and a POSIX
clock on POSIX machines.

boost::chrono is interchangeable with the std::chrono supported in
C++11.

I chose not to upgrade to the newer boost::timer implementation.  The
new boost::timer uses a non-standard duration representation and implies
in the documentation that in future it will upgrade to chrono.