Add Planet::CardinalSuffix
Non-asteroid planets will return the number of non-asteroid planets in
the system that are at a lower orbit, plus itself.  Represented with a
roman numeral.

Asteroid planets will return a localized string and a roman numeral for
the number of asteroid planets in the system at a lower orbit, plus
itself.  If the asteroid planet is the only asteroid in the system, the
roman numeral is not used.