Replace autogenerated effect description with the Effect::Dump call
The autogenerated effect description never produced satisfiying output
in all languages and is mostly used as debugging tool.

The EffectGroup description property can be used to attach more fitting and
better phrased descriptions to specific EffectGroups.

When it comes to debugging the Dump output is more complete and matches
the syntax of the used FOCS scripts better.