Make ListBox::Row Layout persistent.
ListBox::Row::AdjustLayout() deleted and created a new Layout() whenever it
was called.  It was called whenever a ListBox::Row changed, except if
DeferAdjustLayout was used and whenever ListBox::NormalizeRows() was
called.  This caused complicated timing requirements when adjusting
ListBoxes as shown by the code comment in ListBox::SizeMove.

This commit creates a persistent Layout for ListBox::Row which is only
reconstructed in the constructor and clear().

AdjustLayout() and the DeferAdjustLayout class are removed.