Make credit window modal
As reported in #966 on some platforms the credits window didn't capture the
input focus which made it impossible to close the credits.

This commit makes the credit window modal, accepting input and local to
the IntroWnd::OnCredits handler.  Also now the credits window closes on
left mouse button and the escape key.

Fixes #966