Add SaveFileRow::PreRender().
SaveFileRow::PreRender() enables the save dialog to only construct
drawing elements for saved files that are show, while still sorting all
of the files in the directory for which previews have been parsed.

The commit has two parts in order to produce a functional commit:
1. It creates a SaveFileRow::SortKey for each type of SaveFileRow that
sorts based on the appropriate string, not the TextControl which will
not necessarily exist.
2. It creates a Init() function that creates the row columns when they
need to be shown.