Refactor RegisterLinkTags into a free function and use it in HumanClientApp.
RegisterLinkTags was a file scope function called incidentally when a
TextLinker derived class was instantiated.  It registers all TextLinker
known tags with GG::Font so that they can be parsed out of strings.  If
it is not called early enough then string that are expected to be
rendered normally are rendered with their tags embedded.

Making it a free function enables calling it in HumanClientApp before
any text is rendered.