Change SetFirstRowShown to BringRowIntoView in BuildDesignatorWnd.
When a new planet is selected BuildDesignatorWnd uses SetFirstRowShown
to align with the first row with the same drop type as the previously
selected planets first row in the window.  When there are fewer items in
the list than the window size this scrolls some items out of the window,
unnecessarily.

This commit switches to use BringRowIntoView, which shows the whole list
in those circumstances.