Convert file pathes for python open to unicode object
Pythons `open` function differenciate between a regular string parameter
or a unicode string parameter passed to decide on Windows if the should
be passed through the Win32 unicode _wfopen or multibyte sequence API fopen.
The latter is unable to open any path, that isn't encoded in a Window
code page because the byte sequence is not encoded in the native
filesystem encoding.  Using a unicode string object bypasses this
problem.