Update FleetWnd::Refresh() to correctly track fleets.
Add a FleetWnd title that indicates near system X exposes the following
subtle bug in FleetWnd.  When a FleetWnd is created for a travelling
fleet and the fleet arrives at a system the FleetWnd still displays the
fleet and still thinks that it is travelling.

This commit fixes the bug by changing the tracking behavior to the
following.

Are all fleets in one location?  Use that location.

Otherwise, are all selected fleets in one location?  Use that location
and remove fleets not at that location.

Otherwise, is the current location a system?  Use that location and add
appropriate fleets at that system.

Otherwise remove all fleets as all fleets have gone in separate
directions.