AI - add support for automatic history analyser (#1006)
The building is gated by a minimum amount of pp or triggered by a minum number of turns. Both depend on the AI aggression / difficulty setting.

Additional effort is made to make sure the building does not interfere with early game colonization, i.e. the AI should prefer to build colony/outpost ships over this.

The relevant tech is gated by the production scripts that determine whether we want to build this.