Allow industry on outposts (pop center)
Before, all population centers with close to zero population were
depopulated and in turn got the industry reset to zero.
As a outpost is a population center with zero population, it was not
possible to have industry at an outpost.

To be able to build at supply line disconnected outposts, the code now
looks if there is a species on the planet. If it is not, it is considered
an outpost and the industry does not get set to zero.