Fix combat log to not report targeted fighters as Unknown objects.
Fighters were being reported as unknown because they have a temporary
negative ID.  They are the only objects with temporary IDs.

This commit adds the empire associated with targeted objects to the data
stored in a weapon fire event so that temporary ids are reported as
Fighter and not Unknown.