Improve BuildDesignatorWnd tracking of queue state.
Previously, when the BuildDesignator updated it tried to keep the first
row of the queue as the first item of the same type as was in the
original queue.  This caused the queue to jump to either the first
building or the first ship.  Possibly, at some point in the past the
type categories were more specific.

This commit changes the behavior to try and track the position in the
list or keep the same position relative to the end of the list if the
end of the list is visible.

This addresses bug #1147.