Fix drop down list becomes unselectable if app size increases.
Sometimes increasing the application size made drop down lists unusable.

The problem was that the drop down list is a modal window that has an
invisible size equal to the app size when it was instantiated.  The drop
down list could only be within the bounds of the app size when the drop
down list was instantiated.

This commits makes the drop down list track the app size using the
GUI::WindowResizedSignal().

This fixes issue #1139.