Fix bug in SidePanel::DoPanelsLayout.
DoPanelsLayout is unstable when a PlanetPanel that is narrowed for the
scroll bar is also shorter.  This results in the layout flickering
between a SidePanel with vscroll and without.

PlanetPanels with expanded buildings can be shorter with the scroll bar
because it might resize the building icons to be smaller when the scroll
bar is added.

This commit just keeps the layout with the scroll bar even if adding the
scroll makes the panels short enough to not require the scroll bar.