Add states to ClientFSM and ServerFSM to use partial CombatLogs.
Now on start game, load game and the turn update the server informs the
client of all available combat logs.  The clients (only human client)
then downloads the complete logs in the background.

This saves all of the time of uploading combat logs to the AI which
never used them.  This save the time of loading all of the combat logs
since game start in each turn update message.