Fixed inconsistent computation of build number
The Python scripts which compute the build number use the “--format=%h” option of the “git show” command to get the abbreviated SHA hash of the commit at HEAD as the second part of the build number. Apparently, the length of this abbreviated hash has changed with one of the past releases of git, which in turn of course changes the build number computed. To avoid getting slightly different build numbers depending on the git version used, the relevant Python scripts have been fixed to use a (hopefully) more robust and git-version independent way of getting an abbreviated SHA hash.