Fixed inconsistent computation of build number (2nd try)
Git 2.11 changed the default length of the abbreviated SHA-1 hashes which leads to inconsistent results when computing the build number (which the abbreviated SHA-1 of a commit is part of), depending on the git version used. To ensure an equal default length of the abbreviated SHA-1 independent of the git version, the option “--abbrev=7” has been added to the “git show” commands used to acquire the SHA-1, which sets the default length to 7.