In empires.py tighten the exit conditions in HomeSystemFinder.
Fixing the sort order bug found in review #1225 greatly increased the
number and time used in unsuccessful searches for home systems.

This commit adds/tightens 4 early exit conditions:
1. Adds exit early if the minimum limit is achieved.
2. When a succesful candidate is found increase the merit lower bound
one greater than the successful candidate so that the next good
candidate is better.
3. If repeated candidate sets are much smaller than the target, then
exit early since this means that either the min jumps spacing is too
large or the systems pool is too small.
4. If candidate set is close to the correct size, then keep trying.