Add EmpireMeterValue to FOCS for retrieving a value
A generic way to get the value of an empire meter in FOCS.

Use like this:
EmpireMeterValue empire = Source.Owner meter = "METER_IMPERIAL_PP_STORAGE_YIELD"

Be careful that the value only represents the meter value in the instant
when this gets called. E.g. other effects might change it afterwards.

Implementation Note:
Maybe the token should be EmpireMeter instead of EmpireMeterValue.