Fix bug 1237: Monsters not blockading correctly.
This commit fixes UpdateMonsterTravelRestrictions() in ServerApp.cpp.

The problem was that recent code changes caused the loop to loop over
all of the monsters in the galaxy instead of just monsters in one
system.

This commit enforces the following behavior on a per system basis:

1. If any unrestricted (blockading) empire fleets are present, the
monster's  arrival starlane is set to INVALID_OBJECT, so that they can't
leave the system.

2. If any monsters are blockading and any empire fleets are present, the
monster's arrival starlane is set to INVALID_OBJECT, so that they can't
maintain the blockade.